requirements.txt
numpy
torch torchvision
opencv-python
scikit-image
PyYAML